fix: bug with spritesheet generation
